requirements.txt
torch>=1.10.0
numpy>=1.21.0
opencv-python>=4.5.1
matplotlib>=3.4.0
mediapipe>=0.8.7
transformers>=4.5.0
Pillow>=8.0.0